requirements.txt
streamlit
google-generative-ai
python-dotenv
